get orgName from host
