--- analyzer-alert/src/main/java/cloud/erda/analyzer/alert/models/DiceOrg.java
@@ -1,9 +1,0 @@
-package cloud.erda.analyzer.alert.models;
-
-import lombok.Data;
-
-@Data
-public class DiceOrg {
-    private long id;
-    private String name;
-}
